updating to RN 0.29
