+++ android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="com.boundlessgeo.efc"
     android:versionCode="1"
     android:versionName="1.0">
@@ -19,7 +20,7 @@
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
       android:theme="@style/AppTheme"
-      android:name="android.support.multidex.MultiDexApplication">
+      android:name=".MainApplication">
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="AIzaSyBhmo054uWEnyKiWE48hlyTY99fK6DL9hM"/>
