+++ android/app/src/main/res/raw/config.scfg
@@ -1,3 +1,3 @@
 {
-"remote":"http://192.168.0.6:8085"
+"remote":"http://efc-dev.boundlessgeo.com"
 }
