+++ package.json
@@ -7,16 +7,13 @@
     "test": "mocha --compilers js:babel-core/register --require test/setup.js --recursive test/components/*.js"
   },
   "dependencies": {
-    "@turf/distance": "3.3.3",
-    "@turf/inside": "3.3.3",
-    "@turf/point-on-line": "3.3.3",
     "jwt-decode": "2.0.1",
     "react": "^15.1.0",
-    "react-native": "^0.27.2",
+    "react-native": "0.29.2",
     "react-native-button": "^1.6.0",
     "react-native-drawer": "^2.2.3",
     "react-native-image-picker": "0.21.2",
-    "react-native-maps": "0.6.0",
+    "react-native-maps": "file:../react-native-maps",
     "react-native-router-flux": "3.31.0",
     "react-native-slider": "^0.8.0",
     "react-redux": "4.4.5",
